fix(FNP-59): refactored startServer, moved ssr logic to separate route, create isDev helper
