+++ packages/server/utils/index.ts
@@ -1,0 +1,1 @@
+export const isDev = process.env.NODE_ENV === 'development';
